+++ CHANGELOG.md
@@ -9,6 +9,31 @@ template engine project.
 
 ---
 
+## [2.0.0] - 2026-01-25
+
+### Changed
+
+- Change the `get_subblock` method return value to a `Block` object without `None` alternative.
+- Replace the `Block.autotags` attribute with `Block.config.enable_autotags`.
+- Refactor multiple methods.
+- Rewrite documentation to improve structure and clarity.
+
+### Added
+
+- Add support for *Subelements*, i.e., variable and block tags referencing data values defined
+  for nested subblocks.
+- Add support for *variable templates*, i.e., recursive filling of variable and block tags
+  defined in the input data variable values.
+- Add *block variables* defined using the new `Block.config.autotag_blk_var` tag to set the block
+  content into a specified variable instead of keeping it in the original block location.
+- Add automatic indentation of a multiline variable value if the variable tag is indented.
+- Add automatic replacement of a block template in its parent block when a new template is set.
+- Add return flag indicating whether any variable was set or cleared by the `set_variables` and
+  `clear_variables` methods.
+- Add multiple new test functions.
+- Add multiple new demo functions.
+
+
 ## [1.1.0] - 2025-08-04
 
 ### Changed
@@ -64,5 +89,6 @@ below describe modifications since the last
 
 
 [unreleased]: https://github.com/lubomilko/blockie
+[2.0.0]: https://github.com/lubomilko/blockie/releases/tag/2.0.0
 [1.1.0]: https://github.com/lubomilko/blockie/releases/tag/1.1.0
 [1.0.0]: https://github.com/lubomilko/blockie/releases/tag/1.0.0